(FindBug)ボクシングされた値がアンボクシングされて、すぐに再ボクシングされます。com.aimluck.eip.common.ALEipManager.getAclRoleMap(int)
git-svn-id: http://aipo.googlecode.com/svn/trunk@4240 b7d4e767-949e-c8c6-5735-6c11b8a2f74f

+++ core/src/main/java/com/aimluck/eip/common/ALEipManager.java
@@ -372,7 +372,7 @@ public class ALEipManager {
 
     String _featureName;
     for (EipTAclRole _role : roleList) {
-      _featureName = _map.get(_role.getFeatureId().intValue());
+      _featureName = _map.get(_role.getFeatureId());
       roleMap.put(_featureName, _role);
     }
